% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.r
\name{agegenderpieplot}
\alias{agegenderpieplot}
\title{code for plotting age-gender pie plot}
\usage{
agegenderpieplot(charactManufac)
}
\arguments{
\item{charactManufac}{result of charactManufacture code}
}
\description{
code for plotting age-gender pie plot
}
