% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.r
\name{agepieplot}
\alias{agepieplot}
\title{code for plotting age pie plot}
\usage{
agepieplot(charactManufac)
}
\arguments{
\item{charactManufac}{result of charactManufacture code}
}
\description{
code for plotting age pie plot
}
