% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity.r
\name{baseline_comorbidity}
\alias{baseline_comorbidity}
\title{ready for calculate co-prevalence and RR}
\usage{
baseline_comorbidity(connectionDetails, Resultschema, CDMschema,
  cohortTable, cohortId_1, cohortId_2)
}
\description{
ready for calculate co-prevalence and RR
}
