% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicationUse.R
\name{baseline_drug}
\alias{baseline_drug}
\title{ready for calculate co-prevalence and RR}
\usage{
baseline_drug(connectionDetails, Resultschema, CDMschema, cohortTable,
  cohort_definition_id_set)
}
\description{
ready for calculate co-prevalence and RR
}
