% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselineMeasurement.r
\name{baselineMeasure_summary}
\alias{baselineMeasure_summary}
\title{function for calculate mean or median of measurement values}
\usage{
baselineMeasure_summary(x)
}
\description{
function for calculate mean or median of measurement values
}
