% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkerCharacteristic.r
\name{BiomarkerAnalysis}
\alias{BiomarkerAnalysis}
\title{analysis of biomarker characteristics : mean, SD}
\usage{
BiomarkerAnalysis(biomarkerData)
}
\arguments{
\item{biomarkerData}{result of biomarkerManufac}
}
\description{
analysis of biomarker characteristics : mean, SD
}
