% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkerCharacteristic.r
\name{biomarkerManufac}
\alias{biomarkerManufac}
\title{analysis of demographic characteristics according to disease cohort}
\usage{
biomarkerManufac(cohortDefinitionIdSet)
}
\description{
analysis of demographic characteristics according to disease cohort
}
