% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkerCharacteristic.r
\name{biomarkerPvalue}
\alias{biomarkerPvalue}
\title{analysis of biomarker characteristics : p-value if length of cohortDefinitionIdSet is bigger than 2, use ANOVA with Tukey else t.test}
\usage{
biomarkerPvalue(biomarkerData)
}
\arguments{
\item{biomarkerData}{result of biomarkerManufac}
}
\description{
analysis of biomarker characteristics : p-value if length of cohortDefinitionIdSet is bigger than 2, use ANOVA with Tukey else t.test
}
