% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkerCharacteristic.r
\name{biomarkerPvalue_KW}
\alias{biomarkerPvalue_KW}
\title{analysis of biomarker characteristics : p-value using Kruskal Wallis test}
\usage{
biomarkerPvalue_KW(biomarkerData)
}
\arguments{
\item{biomarkerData}{result of biomarkerManufac}
}
\description{
analysis of biomarker characteristics : p-value using Kruskal Wallis test
}
