% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.r
\name{bmitreeplot}
\alias{bmitreeplot}
\title{code for plotting bmi tree plot}
\usage{
bmitreeplot(charactManufac)
}
\arguments{
\item{charactManufac}{result of charactManufacture code}
}
\description{
code for plotting bmi tree plot
}
