% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity.r
\name{calculateRR}
\alias{calculateRR}
\title{ready for calculate Relative Ratio and CI}
\usage{
calculateRR(comorbManufacData, whichDisease)
}
\arguments{
\item{comorbManufacData}{the result of comorbManufacture code (list)}

\item{whichDisease}{metabolic or immune disease}
}
\description{
ready for calculate Relative Ratio and CI
}
