% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalEventRate.r
\name{call_event}
\alias{call_event}
\title{clinical Event data were extracted}
\usage{
call_event(connectionDetails, Resultschema, cohortTable, cohortId_1,
  cohortId_2, cohortId_event)
}
\arguments{
\item{connectionDetails}{}

\item{Resultschema}{}

\item{cohortId_1}{}

\item{cohortId_2}{}

\item{cohortId_event}{}
}
\description{
clinical Event data were extracted
}
