% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalCharacteristic.r
\name{characPvalue}
\alias{characPvalue}
\title{analysis of demographic characteristics : p-value using ANOVA with Tukey}
\usage{
characPvalue(clinicalCharData)
}
\arguments{
\item{clinicalCharData}{result of clinicalCharManufacture}
}
\description{
analysis of demographic characteristics : p-value using ANOVA with Tukey
}
