% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalCharacteristic.r
\name{characterAnalysis}
\alias{characterAnalysis}
\title{analysis of demographic characteristics : mean, SD}
\usage{
characterAnalysis(clinicalCharData)
}
\arguments{
\item{clinicalCharData}{result of clinicalCharManufacture}
}
\description{
analysis of demographic characteristics : mean, SD
}
