% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalCharacteristic.r
\name{characteristic_pvalue}
\alias{characteristic_pvalue}
\title{analysis of demographic characteristics : if the value has normality, p value is calculated by t-test, if not by wilcoxon in continuous value}
\usage{
characteristic_pvalue(characteristic_manufac)
}
\arguments{
\item{characteristic_manufac}{result of clinicalCharManufacture}
}
\description{
analysis of demographic characteristics : if the value has normality, p value is calculated by t-test, if not by wilcoxon in continuous value
}
