% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalCharacteristic.r
\name{characteristic_summary}
\alias{characteristic_summary}
\title{analysis of demographic characteristics; if the value has normality, calculate mean +/- sd, if not median (25\%, 75\% quantile) in continuous value}

\usage{
characteristic_summary(characteristic_manufac)
}
\arguments{
\item{characteristic_manufac}{result of charterstic_manufacture}
}
\description{
analysis of demographic characteristics; if the value has normality, calculate mean +/- sd, if not median (25\%, 75\% quantile) in continuous value
}
