% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.r
\name{charactManufacture}
\alias{charactManufacture}
\title{analysis of demographic characteristics according to disease cohort}
\usage{
charactManufacture(characteristicData, cohortDefinitionIdSet)
}
\description{
analysis of demographic characteristics according to disease cohort
}
