% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalCharacteristic.r
\name{charterstic_manufacture}
\alias{charterstic_manufacture}
\title{collect and mutate demographic characteristics}
\usage{
charterstic_manufacture(cohort_definition_id_set)
}
\description{
collect and mutate demographic characteristics
}
