% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalCharacteristic.r
\name{clinicalCharManufacture}
\alias{clinicalCharManufacture}
\title{analysis of demographic characteristics according to disease cohort}
\usage{
clinicalCharManufacture(cohortDefinitionIdSet)
}
\description{
analysis of demographic characteristics according to disease cohort
}
