% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity.r
\name{co_prevtable}
\alias{co_prevtable}
\title{calculate co-prevalence rate}
\usage{
co_prevtable(comorbManufacData, whichDisease)
}
\arguments{
\item{comorbManufacData}{the result of comorbManufacture code (list)}

\item{whichDisease}{metabolic or immune disease}
}
\description{
calculate co-prevalence rate
}
