% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity.r
\name{comorbManufacture}
\alias{comorbManufacture}
\title{ready for plotting comorbidity data}
\usage{
comorbManufacture(comorbidityData, cohortDefinitionIdSet)
}
\description{
ready for plotting comorbidity data
}
