% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.r
\name{demographic_Table}
\alias{demographic_Table}
\title{table output of bmi and gender proportion}
\usage{
demographic_Table(charactManufac, dividedVariable)
}
\arguments{
\item{charactManufac}{result of charactManufacture code}

\item{dividedVariable}{age, bmi (gender was default)}
}
\description{
table output of bmi and gender proportion
}
