% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalEventRate.r
\name{event_incidence}
\alias{event_incidence}
\title{combine and calculate incidence rate of clinical event}
\usage{
event_incidence(callEvent_result)
}
\arguments{
\item{callEvent_result}{}
}
\description{
combine and calculate incidence rate of clinical event
}
