% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exacerbationCount.r
\name{exacerbaManufacture}
\alias{exacerbaManufacture}
\title{exacerbation count according to asthma cohort}
\usage{
exacerbaManufacture()
}
\description{
exacerbation count according to asthma cohort
}
