% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.r
\name{genderpieplot}
\alias{genderpieplot}
\title{code for plotting gender Pie plot}
\usage{
genderpieplot(charactManufac)
}
\arguments{
\item{charactManufac}{result of charactManufacture code}
}
\description{
code for plotting gender Pie plot
}
