% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal.r
\name{getAllLongitudinal}
\alias{getAllLongitudinal}
\title{extract longitudinal analysis using FeatureExtraction}
\usage{
getAllLongitudinal(connectionDetails, CDMschema, Resultschema, cohortTable,
  cohortId)
}
\arguments{
\item{connectionDetails}{}

\item{cohortTable}{}

\item{cohortId}{}

\item{cdmDatabaseSchema}{}

\item{resultsDatabaseSchema}{}
}
\description{
extract longitudinal analysis using FeatureExtraction
}
