% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselineMeasurement.r
\name{getBaselineCovariate}
\alias{getBaselineCovariate}
\title{get measurement baseline measurement data}
\usage{
getBaselineCovariate(connectionDetails, CDMschema, Resultschema,
  cohortTable, cohortId)
}
\description{
get measurement baseline measurement data
}
