% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity.r
\name{getConditionCovariate}
\alias{getConditionCovariate}
\title{ready for plotting comorbidity data}
\usage{
getConditionCovariate(connectionDetails, Resultschema, CDMschema,
  cohortTable, cohortId)
}
\description{
ready for plotting comorbidity data
}
