% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicationUse.R
\name{getDrugCovariate}
\alias{getDrugCovariate}
\title{Analysis of baseline medication use of cohorts}
\usage{
getDrugCovariate(connectionDetails, Resultschema, CDMschema, cohortTable,
  cohortId)
}
\description{
Analysis of baseline medication use of cohorts
}
