% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal.r
\name{getLongitudinal}
\alias{getLongitudinal}
\title{subsetting features that have specific measurement_concept_id}
\usage{
getLongitudinal(all_longitudinal_data, measurement_concept_id,
  time_unit = "year")
}
\arguments{
\item{all_longitudinal_data}{}

\item{measurement_concept_id}{}

\item{time_unit}{}
}
\description{
subsetting features that have specific measurement_concept_id
}
