% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictiveVariables.r
\name{getPlpData}
\alias{getPlpData}
\title{create Plp data and population data (ready for run Plp)}
\usage{
getPlpData(connectionDetails, connection, Resultschema, CDMschema,
  cohortTable = "asthma_cohort", targetId = 1, outcomeCohortConceptId,
  covariateSetting, washoutPeriod = 0,
  removeSubjectsWithPriorOutcome = TRUE, riskWindowStart = 1,
  riskWindowEnd = 365 * 15)
}
\arguments{
\item{connectionDetails}{}

\item{connection}{}

\item{Resultschema}{}

\item{CDMschema}{}

\item{cohortTable}{table name which contains asthma_cohort (default = 'asthma_cohort')}

\item{targetId}{target cohort Id (default = 1)}

\item{outcomeCohortConceptId}{outcome cohort Id (2 = NSA, 3 = SA, 4 = AERD, 5 = ATA)}

\item{covariateSetting}{}

\item{washoutPeriod}{0}

\item{removeSubjectsWithPriorOutcome}{TRUE}

\item{riskWindowStart}{1}

\item{riskWindowEnd}{365*15}
}
\description{
create Plp data and population data (ready for run Plp)
}
