% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertLCMMResultCohort.r
\name{insertCohort}
\alias{insertCohort}
\title{code for insert cohort at result table}
\usage{
insertCohort(newCohortIdSet, resultOflcmm, connectionDetails, Resultschema,
  cohortTable)
}
\description{
code for insert cohort at result table
}
