% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMMclustering.r
\name{latent_class_classification}
\alias{latent_class_classification}
\title{using latent class linear mixed model (LCMM), create groups according to longitudinally measured data trajectories}
\usage{
latent_class_classification(longitudinal_data, cluster_number = 3)
}
\arguments{
\item{longitudinal_data}{result of PFTmanufacture function}

\item{cluster_number}{}
}
\description{
using latent class linear mixed model (LCMM), create groups according to longitudinally measured data trajectories
}
