% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMMclustering.r
\name{latent_longitudinal_plot}
\alias{latent_longitudinal_plot}
\title{trajectory plot clustered by lcmm}
\usage{
latent_longitudinal_plot(lcmm_out, cluster_number)
}
\arguments{
\item{lcmm_out}{result of PFTmanufacture function}

\item{cluster_number}{}
}
\description{
trajectory plot clustered by lcmm
}
