% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMMclustering.r
\name{latent_longitudinal_table}
\alias{latent_longitudinal_table}
\title{trajectory table clustered by lcmm}
\usage{
latent_longitudinal_table(lcmm_classification_result_list, cluster_number)
}
\arguments{
\item{lcmm_classification_result_list}{result of latent_class_classification function}

\item{cluster_number}{the number of clusters}
}
\description{
trajectory table clustered by lcmm
}
