% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal.r
\name{lme_logitudinal}
\alias{lme_logitudinal}
\title{use linear mixed model regression, find regression line and its CI 95\%}
\usage{
lme_logitudinal(longitudinalData)
}
\arguments{
\item{longitudinalData}{result of PFTmanufacture function}
}
\description{
use linear mixed model regression, find regression line and its CI 95\%
}
