% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal.r
\name{lmePft}
\alias{lmePft}
\title{use linear mixed model regression, find regression line and its CI 95%}
\usage{
lmePft(longitudinalData)
}
\arguments{
\item{longitudinalData}{result of PFTmanufacture function}
}
\description{
use linear mixed model regression, find regression line and its CI 95%
}
