% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal.r
\name{longitudinal}
\alias{longitudinal}
\title{aggregate total code for longitudinal analysis}
\usage{
longitudinal(connectionDetails, CDMschema, Resultschema, cohortTable,
  cohortId_1, cohortId_2, measurement_concept_id)
}
\description{
aggregate total code for longitudinal analysis
}
