% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exacerbationCount.r
\name{p_value_ExacerbationCouny}
\alias{p_value_ExacerbationCouny}
\title{t-test p-value calculate about exacerbation count}
\usage{
p_value_ExacerbationCouny(exacerbationCount, cohortDefinitionIdSet)
}
\arguments{
\item{exacerbationCount}{the result of exacerbaManufacture function}
}
\description{
t-test p-value calculate about exacerbation count
}
