% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulmonaryfunction.r
\name{pft_count_table}
\alias{pft_count_table}
\title{calculate count of cohort population}
\usage{
pft_count_table(PFTmanufactured)
}
\arguments{
\item{PFTmanufactured}{result of PFTmanufacture code}
}
\description{
calculate count of cohort population
}
