% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulmonaryfunction.r
\name{pft_predict_table}
\alias{pft_predict_table}
\title{calculate prediction values of PFT that 5, 10, 15 years after treatment}
\usage{
pft_predict_table(PFTmanufactured)
}
\arguments{
\item{PFTmanufactured}{result of PFTmanufacture code}
}
\description{
calculate prediction values of PFT that 5, 10, 15 years after treatment
}
