% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulmonaryfunctionInDetail.r
\name{pftCountTable_indetail}
\alias{pftCountTable_indetail}
\title{calculate count of cohort population in detail}
\usage{
pftCountTable_indetail(PFTmanufactured, genderDivided)
}
\arguments{
\item{PFTmanufactured}{result of PFTmanufacture code}

\item{genderDivided}{logical (TRUE/FALSE)}
}
\description{
calculate count of cohort population in detail
}
