% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulmonaryfunction.r
\name{PFTmanufacture}
\alias{PFTmanufacture}
\title{ready for plotting PFT data}
\usage{
PFTmanufacture(measurementData, measurementType, cohortDefinitionIdSet)
}
\arguments{
\item{measurementConceptId}{FEV1/FVC(%) = 3011505, FEV1(%) = 3011708}
}
\description{
ready for plotting PFT data
}
