% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulmonaryfunctionInDetail.r
\name{PFTmanufacture_detail}
\alias{PFTmanufacture_detail}
\title{ready for PFT data getting clear in detail}
\usage{
PFTmanufacture_detail(measurementType, cohortDefinitionIdSet,
  ageSection = NULL)
}
\arguments{
\item{ageSection}{character Input separated by '/'}

\item{measurementConceptId}{FEV1/FVC(%) = 3011505, FEV1(%) = 3011708}
}
\description{
ready for PFT data getting clear in detail
}
