% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulmonaryfunctionInDetail.r
\name{pftPredictTable_indetail}
\alias{pftPredictTable_indetail}
\title{calculate prediction values of PFT that 5, 10, 15 years after treatment in detail}
\usage{
pftPredictTable_indetail(PFTmanufactured, genderDivided)
}
\arguments{
\item{PFTmanufactured}{result of PFTmanufacture code}

\item{genderDivided}{logical (TRUE/FALSE)}
}
\description{
calculate prediction values of PFT that 5, 10, 15 years after treatment in detail
}
