% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalEventRate.r
\name{plot_event_rate}
\alias{plot_event_rate}
\title{plot for clinical event rate}
\usage{
plot_event_rate(event_result)
}
\arguments{
\item{event_result}{}
}
\description{
plot for clinical event rate
}
