% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exacerbationCount.r
\name{plotExacerbationCount}
\alias{plotExacerbationCount}
\title{plot exacerbation Count}
\usage{
plotExacerbationCount(sumCountExacerbation)
}
\arguments{
\item{sumCountExacerbation}{the result of sumCountExacerbation function}
}
\description{
plot exacerbation Count
}
