% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal.r
\name{plotLmm}
\alias{plotLmm}
\title{plot trajectory line and its CI}
\usage{
plotLmm(longitudinalData, lmeData, pftIndividual = TRUE,
  ConfidencialIntervalVisualize = TRUE)
}
\arguments{
\item{pftIndividual}{logical (TRUE/FALSE)}

\item{ConfidencialIntervalVisualize}{logical (TRUE/FALSE)}
}
\description{
plot trajectory line and its CI
}
