% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulmonaryfunction.r
\name{plotPFT}
\alias{plotPFT}
\title{plot PFT according to time flow}
\usage{
plotPFT(PFTmanufactured)
}
\arguments{
\item{PFTmanufactured}{result of PFTmanufacture code}
}
\description{
plot PFT according to time flow
}
