% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulmonaryfunctionInDetail.r
\name{plotPFT_detail}
\alias{plotPFT_detail}
\title{plot PFT according to time flow in detail (add more covariate)}
\usage{
plotPFT_detail(PFTmanufactured, genderDivided)
}
\arguments{
\item{PFTmanufactured}{result of PFTmanufacture code}

\item{genderDivided}{logical (TRUE/FALSE)}
}
\description{
plot PFT according to time flow in detail (add more covariate)
}
