% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal.r
\name{plotpftLmm}
\alias{plotpftLmm}
\title{plot pft trajectory line and its CI}
\usage{
plotpftLmm(longitudinalData, lmePftData, pftIndividual = TRUE,
  ConfidencialIntervalVisualize = TRUE)
}
\arguments{
\item{pftIndividual}{logical (TRUE/FALSE)}

\item{ConfidencialIntervalVisualize}{logical (TRUE/FALSE)}
}
\description{
plot pft trajectory line and its CI
}
