% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalCharacteristic.r
\name{pvalueCalAmongMoreThanTwo}
\alias{pvalueCalAmongMoreThanTwo}
\title{calculate pvalue among more than two cohorts}
\usage{
pvalueCalAmongMoreThanTwo(y, index, cohortIdLength)
}
\arguments{
\item{y}{must be vector of numeric data}
}
\description{
calculate pvalue among more than two cohorts
}
