% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalCharacteristic.r
\name{pvalueCalBetweenTwo}
\alias{pvalueCalBetweenTwo}
\title{calculate pvalue between two cohorts : if the value has normality, p value is calculated by t-test, if not by wilcoxon in continuous value}
\usage{
pvalueCalBetweenTwo(y, index, cohortIdLength)
}
\arguments{
\item{y}{must be vector of numeric data}
}
\description{
calculate pvalue between two cohorts : if the value has normality, p value is calculated by t-test, if not by wilcoxon in continuous value
}
