% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictiveVariables.r
\name{RunPlp}
\alias{RunPlp}
\title{Run Plp}
\usage{
RunPlp(getplpOut, learningModel)
}
\arguments{
\item{getplpOut}{get from getPlpData code}

\item{learningModel}{machin learning model to develop prediction model (lassologistic, gradientboosting, randomforest)}
}
\description{
Run Plp
}
