% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting.r
\name{setting}
\alias{setting}
\title{Setting some list or dataframe (diseaseList for comorbidity analysis, covariateSetting for PLP analysis)}
\usage{
setting()
}
\description{
Setting some list or dataframe (diseaseList for comorbidity analysis, covariateSetting for PLP analysis)
}
