% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_data.r
\name{sql_reader}
\alias{sql_reader}
\title{code for reading query and calling the data (using SqlRender)}
\usage{
sql_reader(sqlquery, Resultschema, CDMschema, connectionDetails,
  connection)
}
\description{
code for reading query and calling the data (using SqlRender)
}
