% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exacerbationCount.r
\name{sumCountExacerbation}
\alias{sumCountExacerbation}
\title{sum the exacerbation count according to asthma cohort}
\usage{
sumCountExacerbation(exacerbationCount, cohortDefinitionIdSet)
}
\arguments{
\item{exacerbationCount}{the result of exacerbaManufacture function}
}
\description{
sum the exacerbation count according to asthma cohort
}
