% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helper.r
\name{switchselect_model}
\alias{switchselect_model}
\title{switch input from label to machine learning model}
\usage{
switchselect_model(input)
}
\description{
switch input from label to machine learning model
}
