% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helper.r
\name{switchselect_plp}
\alias{switchselect_plp}
\title{switch input from label to cohortDefinitionId for plp}
\usage{
switchselect_plp(input)
}
\description{
switch input from label to cohortDefinitionId for plp
}
