% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal.r
\name{tableLmm}
\alias{tableLmm}
\title{table for representative estimated profile trajectory and its CI}
\usage{
tableLmm(longitudinal_result)
}
\arguments{
\item{longitudinal_result}{result of longitudinal code}
}
\description{
table for representative estimated profile trajectory and its CI
}
