% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictiveVariables.r
\name{tablePredictiveVariables}
\alias{tablePredictiveVariables}
\title{ranking top n predictive variables and conceptName}
\usage{
tablePredictiveVariables(machineLearningData, rankCount = 20)
}
\arguments{
\item{machineLearningData}{get from RunPlp code}

\item{rankCount}{how many variables do you want to see?}
}
\description{
ranking top n predictive variables and conceptName
}
