% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselineMeasurement.r
\name{baselineMeasure_compare}
\alias{baselineMeasure_compare}
\title{comparing measurement value between two cohorts}
\usage{
baselineMeasure_compare(connectionDetails, Resultschema, CDMschema,
  cohortTable, cohort_definition_id_set, measurementConceptIdSet = c(2,
  3, 5, 6, 7, 8, 3028930, 4169578, 44786758, 4010492, 3046594, 2212469,
  3011708, 3006504, 3005322, 3005600, 3017501, 3026514, 3005791, 3021940,
  3011505, 3013115, 3018010, 3022096))
}
\arguments{
\item{connectionDetails}{}

\item{Resultschema}{}

\item{CDMschema}{}

\item{cohortTable}{}

\item{cohort_definition_id_set}{}

\item{measurementConceptIdSet}{}
}
\description{
comparing measurement value between two cohorts
}
