% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselineMeasurement.r
\name{baselineMeasure_pvalue}
\alias{baselineMeasure_pvalue}
\title{function for calculate p-value of measurement values}
\usage{
baselineMeasure_pvalue(measuredData)
}
\description{
function for calculate p-value of measurement values
}
