% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarkerCharacteristic.r
\name{biomarkerPvalue}
\alias{biomarkerPvalue}
\title{analysis of biomarker characteristics : p-value using ANOVA with Tukey}
\usage{
biomarkerPvalue(biomarkerData)
}
\arguments{
\item{biomarkerData}{result of biomarkerManufac}
}
\description{
analysis of biomarker characteristics : p-value using ANOVA with Tukey
}
