% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicalEventRate.r
\name{call_event}
\alias{call_event}
\title{clinical Event data were extracted}
\usage{
call_event(cohort_definition_id_set, cohortId_event)
}
\arguments{
\item{cohort_definition_id_set}{}

\item{cohortId_event}{}
}
\description{
clinical Event data were extracted
}
