% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.r
\name{demographic_cal}
\alias{demographic_cal}
\title{calculate p-value of gender(chi-square test) and and age mean(t-test)}
\usage{
demographic_cal(charactManufac, dividedVariable)
}
\arguments{
\item{charactManufac}{result of charactManufacture code}

\item{dividedVariable}{age, bmi (gender was default)}
}
\description{
calculate p-value of gender(chi-square test) and and age mean(t-test)
}
