% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFreeSurvival.r
\name{eventFreeSurvival}
\alias{eventFreeSurvival}
\title{get measurement baseline measurement data}
\usage{
eventFreeSurvival(cohort_definition_id_set, cohortId_event,
  targetSurvivalEndDate)
}
\description{
get measurement baseline measurement data
}
