% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictiveVariables.r
\name{getPlpData}
\alias{getPlpData}
\title{create Plp data and population data (ready for run Plp)}
\usage{
getPlpData(connectionDetails, connection, Resultschema, CDMschema,
  cohortTable = "asthma_cohort", targetCohortConceptId,
  outcomeCohortConceptId, covariateSetting, washoutPeriod = 0,
  removeSubjectsWithPriorOutcome = TRUE, riskWindowStart, riskWindowEnd,
  minTimeAtRisk)
}
\arguments{
\item{connectionDetails}{}

\item{connection}{}

\item{Resultschema}{}

\item{CDMschema}{}

\item{cohortTable}{table name which contains asthma_cohort (default = 'asthma_cohort')}

\item{outcomeCohortConceptId}{outcome cohort Id}

\item{covariateSetting}{}

\item{washoutPeriod}{0}

\item{removeSubjectsWithPriorOutcome}{TRUE}

\item{riskWindowStart}{}

\item{riskWindowEnd}{}

\item{minTimeAtRisk}{}

\item{targetId}{target cohort Id (default = 1)}
}
\description{
create Plp data and population data (ready for run Plp)
}
