% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertLCMMResultCohort.r
\name{insertCohort}
\alias{insertCohort}
\title{code for insert cohort at result table}
\usage{
insertCohort(newCohortIdSet, target_cluster_cohort, resultOflcmm,
  connection, Resultschema, cohortTable)
}
\arguments{
\item{newCohortIdSet}{}

\item{target_cluster_cohort}{}

\item{resultOflcmm}{results of latent_class_classification code}
}
\description{
code for insert cohort at result table
}
