% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMMclustering.r
\name{latent_longitudinal_plot}
\alias{latent_longitudinal_plot}
\title{trajectory plot clustered by lcmm}
\usage{
latent_longitudinal_plot(lcmm_classification_result_list,
  individual_trajectories = TRUE, cluster_number)
}
\arguments{
\item{lcmm_classification_result_list}{result of latent_class_classification function}

\item{individual_trajectories}{logical value (TRUE/FALSE) ; being individual trajectories?}

\item{cluster_number}{the number of clusters}
}
\description{
trajectory plot clustered by lcmm
}
