% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal.r
\name{longitudinal}
\alias{longitudinal}
\title{aggregate total code for longitudinal analysis}
\usage{
longitudinal(allLongitudinal_cohort_1, allLongitudinal_cohort_2,
  measurement_concept_id)
}
\description{
aggregate total code for longitudinal analysis
}
