% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal.r
\name{plotLmm}
\alias{plotLmm}
\title{plot trajectory line and its CI}
\usage{
plotLmm(longitudinal_result, pftIndividual = TRUE)
}
\arguments{
\item{longitudinal_result}{result of longitudinal code}

\item{pftIndividual}{logical (TRUE/FALSE)}
}
\description{
plot trajectory line and its CI
}
