% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictiveVariables.r
\name{RunPlp}
\alias{RunPlp}
\title{Run Plp}
\usage{
RunPlp(getplpOut, learningModel, splitSeed = NULL,
  outputFolder = outputFolder)
}
\arguments{
\item{getplpOut}{get from getPlpData code}

\item{learningModel}{input$ModelSelect}

\item{splitSeed}{seed setting}
}
\description{
Run Plp
}
