% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exacerbationCount.r
\name{tableExacerbationCoun}
\alias{tableExacerbationCoun}
\title{table of exacerbation Count}
\usage{
tableExacerbationCoun(sumCountExacerbation)
}
\arguments{
\item{sumCountExacerbation}{the result of sumCountExacerbation function}
}
\description{
table of exacerbation Count
}
